requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
atomicwrites==1.4.0
attrs==21.2.0
backports.functools-lru-cache==1.6.4
certifi==2021.5.30
charset-normalizer==2.0.3
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
future==0.18.2
idna==3.2
iniconfig==1.1.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
requests==2.26.0
selene==2.0.0a38
selenium==3.141.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
webdriver-manager==3.4.1
